# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FitTrack API — fitness tracking, nutrition logging, and social features
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /exercises:
    get:
      operationId: listExercises
      tags: [exercises]
      summary: List exercises (optionally filter by muscle group or search)
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: muscleGroup
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Exercise list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
    post:
      operationId: createExercise
      tags: [exercises]
      summary: Create a custom exercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExerciseInput"
      responses:
        "201":
          description: Created exercise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
  /exercises/{exerciseId}:
    get:
      operationId: getExercise
      tags: [exercises]
      summary: Get exercise details and PR history
      parameters:
        - name: exerciseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exercise detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
  /exercises/{exerciseId}/history:
    get:
      operationId: getExerciseHistory
      tags: [exercises]
      summary: Get set history for an exercise (for progress chart)
      parameters:
        - name: exerciseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: History data points
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseHistoryResult"

  # ─── Routines ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Exercise:
      type: object
      required: [id, name, category]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        muscleGroups:
          type: array
          items:
            type: string
        description:
          type: ["string", "null"]
        isCustom:
          type: boolean
        personalRecord:
          type: ["number", "null"]
    ExerciseInput:
      type: object
      required: [name, category]
      properties:
        name:
          type: string
        category:
          type: string
        muscleGroups:
          type: array
          items:
            type: string
        description:
          type: string
    ExerciseHistoryPoint:
      type: object
      required: [date, value]
      properties:
        date:
          type: string
          format: date
        value:
          type: number
        workoutId:
          type: integer
    ExerciseHistoryResult:
      type: object
      required: [weight, oneRepMax, volume]
      properties:
        weight:
          type: array
          items:
            $ref: "#/components/schemas/ExerciseHistoryPoint"
        oneRepMax:
          type: array
          items:
            $ref: "#/components/schemas/ExerciseHistoryPoint"
        volume:
          type: array
          items:
            $ref: "#/components/schemas/ExerciseHistoryPoint"
        personalRecord:
          type: ["number", "null"]

    # ─── Routines ──────────────────────────────────────────────────────────────
...